Enable treat warnings as errors on msvc

+++ src/platform/win_x86_64/_write.c
@@ -4,12 +4,12 @@
 
 #include "file.h"
 
-#include <windows.h>
+#include "windows_workaround.h"
 
 int _Write(FILE *fd, const void *buf, size_t count) {
     DWORD written;
 
-    if (!WriteFile(fd->handle, buf, count, &written, NULL)) {
+    if (!WriteFile(fd->handle, buf, (DWORD)count, &written, NULL)) {
         return -1;
     }
